+++ changelog.md
@@ -5,6 +5,7 @@ v1.16.0:
 - Add initial compatibility with WoW 12.x / Midnight. Please see readme/in-game FAQ page for details about this significant change.
 - Add option to forcibly disable cooldown numbers on TargetInfo buff and debuff as well as class power frames (DK runes, Paladin holy power, etc.) when set to "Cooldown" or "Both" mode.
 - Add option to hide the default breath ("mirror") bar when the IceHUD MirrorBar is enabled (the built-in bar is now hidden by default).
+- Fix Vigor display for Druid flight form in 11.2.7.
 
 v1.15.23: